SInput Implementation

+++ main/include/interval.h
@@ -4,8 +4,7 @@
 #include <stdint.h>
 #include "hoja_types.h"
 
-bool interval_run(uint32_t timestamp, uint32_t interval, interval_s *state);
-
-bool interval_resettable_run(uint32_t timestamp, uint32_t interval, bool reset, interval_s *state);
+bool interval_run(uint64_t timestamp, uint64_t interval, interval_s *state);
+bool interval_resettable_run(uint64_t timestamp, uint64_t interval, bool reset, interval_s *state);
 
 #endif
